pyproject.toml
[project]
name = "lg-quickstart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipython>=9.3.0",
    "langchain[anthropic,openai]>=0.3.26",
    "langgraph>=0.4.10",
    "langsmith>=0.4.2",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.1",
]
